requirements/staging.txt
-r base.txt
dj-database-url==0.2.1
psycopg2==2.7.3.2
boto3
django-storages